feat:테스트환경 위한 vdb세팅 수정
