--- app-tuning/core/vector_database/__init__.py
@@ -1,26 +1,0 @@
-from .client import get_chroma_client
-from .collections import (
-    get_similarity_collection,
-    get_user_collection,
-    reset_collections,
-)
-from .similarity_repository import (
-    clean_up_similarity,
-    get_user_similarities,
-    list_similarities,
-)
-from .user_repository import delete_user, get_user_data, get_users_data, list_users
-
-__all__ = [
-    "get_chroma_client",
-    "get_similarity_collection",
-    "get_user_collection",
-    "reset_collections",
-    "clean_up_similarity",
-    "get_user_similarities",
-    "list_similarities",
-    "delete_user",
-    "get_user_data",
-    "get_users_data",
-    "list_users",
-]
